Develop

+++ Anaface/TScrollerControl.cpp
@@ -182,7 +182,7 @@ TrecPointer<TControl> TScrollerControl::GetChildControl()
 {
 	return childControl;
 }
-bool TScrollerControl::onScroll(int x, int y)
+bool TScrollerControl::onScroll(float x,float y)
 {
 	if (childControl.Get())
 	{
